pyproject.toml
[project]
name = "investor-graph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grandalf>=0.8",
    "langchain>=0.3.27",
    "langchain-fireworks>=0.3.0",
    "langchain-huggingface>=0.3.1",
    "langgraph>=0.6.8",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
